+++ adapters/I-Mobile/IMobileAdapter/GADMAdapterIMobileConstants.h
@@ -1,4 +1,4 @@
-// Copyright 2019 Google Inc.
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,6 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+#import <Foundation/Foundation.h>
 
 /// Adapter version.
 static NSString *const kGADMAdapterIMobileVersion = @"2.0.33.0";