+++ adapters/I-Mobile/IMobileAdapter/GADMediationAdapterIMobile.m
@@ -1,4 +1,4 @@
-// Copyright 2019 Google Inc.
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,28 +13,28 @@
 // limitations under the License.
 
 #import "GADMediationAdapterIMobile.h"
-#import "GADIMobileMediatedUnifiedNativeAd.h"
+
+#import "GADMAdapterIMobileBannerAd.h"
 #import "GADMAdapterIMobileConstants.h"
+#import "GADMAdapterIMobileInterstitialAd.h"
+#import "GADMAdapterIMobileUnifiedNativeAd.h"
 #import "GADMAdapterIMobileUtils.h"
 
+@interface GADMediationAdapterIMobile () <GADMediationAdapter>
+@end
+
 @implementation GADMediationAdapterIMobile {
-  /// Connector for AdMob.
-  __weak id<GADMAdNetworkConnector> _connector;
+  /// i-mobile banner ad wrapper.
+  GADMAdapterIMobileBannerAd *_bannerAd;
 
-  /// View for i-mobile SDK.
-  UIView *_sdkView;
+  /// i-mobile interstitial ad wrapper.
+  GADMAdapterIMobileInterstitialAd *_interstitialAd;
 
-  /// i-mobile spot id.
-  NSString *_spotID;
+  /// i-mobile native ad wrapper.
+  GADMAdapterIMobileUnifiedNativeAd *_unifiedNativeAd;
 }
 
 #pragma mark - GADMediationAdapter
-
-+ (void)setUpWithConfiguration:(GADMediationServerConfiguration *)configuration
-             completionHandler:(GADMediationAdapterSetUpCompletionBlock)completionHandler {
-  // i-Mobile SDK doesn't have any initialization API.
-  completionHandler(nil);
-}
 
 + (GADVersionNumber)adSDKVersion {
   // i-Mobile SDK doesn't have any API to get the version.
@@ -43,6 +43,10 @@
 }
 
 + (GADVersionNumber)version {
+  return [GADMediationAdapterIMobile adapterVersion];
+}
+
++ (GADVersionNumber)adapterVersion {
   GADVersionNumber version = {0};
   NSArray<NSString *> *components = [kGADMAdapterIMobileVersion componentsSeparatedByString:@"."];
 
@@ -55,119 +59,36 @@
   return version;
 }
 
-#pragma mark - GADMAdNetworkAdapter
-
-+ (NSString *)adapterVersion {
-  return kGADMAdapterIMobileVersion;
-}
-
 + (Class<GADAdNetworkExtras>)networkExtrasClass {
   return Nil;
 }
 
-- (instancetype)initWithGADMAdNetworkConnector:(id<GADMAdNetworkConnector>)connector {
-  self = [super init];
-  if (self) {
-    // Initialize.
-    _connector = connector;
-    _sdkView = [[UIView alloc] init];
-
-    // Get parameters for i-mobile SDK.
-    NSDictionary<NSString *, NSString *> *params = [connector credentials];
-    NSString *publisherId = params[kGADMAdapterIMobilePublisherIdKey];
-    NSString *mediaId = params[kGADMAdapterIMobileMediaIdKey];
-    _spotID = params[kGADMAdapterIMobileSpotIdKey];
-
-    // Call i-mobile SDK.
-    [ImobileSdkAds registerWithPublisherID:publisherId MediaID:mediaId SpotID:_spotID];
-    [ImobileSdkAds startBySpotID:_spotID];
-  }
-  return self;
++ (void)setUpWithConfiguration:(GADMediationServerConfiguration *)configuration
+             completionHandler:(GADMediationAdapterSetUpCompletionBlock)completionHandler {
+  // i-Mobile SDK doesn't have any initialization API.
+  completionHandler(nil);
 }
 
-/// Not supported.
-- (void)getBannerWithSize:(GADAdSize)adSize {
-  NSString *errorString =
-      @"GADMediationAdapterIMobile doesn't support banner ads. Please use GADMAdapterIMobile.";
-  NSError *error =
-      GADMAdapterIMobileErrorWithCodeAndDescription(kGADErrorInvalidRequest, errorString);
-  [_connector adapter:self didFailAd:error];
+- (void)loadBannerForAdConfiguration:(GADMediationBannerAdConfiguration *)adConfiguration
+                   completionHandler:(GADMediationBannerLoadCompletionHandler)completionHandler {
+  _bannerAd = [[GADMAdapterIMobileBannerAd alloc] initWithAdConfiguration:adConfiguration];
+  [_bannerAd loadBannerAdWithCompletionHandler:completionHandler];
 }
 
-/// Not supported.
-- (void)getInterstitial {
-  NSString *errorString = @"GADMediationAdapterIMobile doesn't support interstitial ads. Please "
-                          @"use GADMAdapterIMobile.";
-  NSError *error =
-      GADMAdapterIMobileErrorWithCodeAndDescription(kGADErrorInvalidRequest, errorString);
-  [_connector adapter:self didFailAd:error];
+- (void)loadInterstitialForAdConfiguration:
+            (GADMediationInterstitialAdConfiguration *)adConfiguration
+                         completionHandler:
+                             (GADMediationInterstitialLoadCompletionHandler)completionHandler {
+  _interstitialAd =
+      [[GADMAdapterIMobileInterstitialAd alloc] initWithAdConfiguration:adConfiguration];
+  [_interstitialAd loadInterstitialAdWithCompletionHandler:completionHandler];
 }
 
-- (void)stopBeingDelegate {
-  _sdkView = nil;
-}
-
-/// Not supported.
-- (void)presentInterstitialFromRootViewController:(UIViewController *)rootViewController {
-  NSString *errorString = @"GADMediationAdapterIMobile doesn't support interstitial ads. Please "
-                          @"use GADMAdapterIMobile.";
-  NSError *error =
-      GADMAdapterIMobileErrorWithCodeAndDescription(kGADErrorInvalidRequest, errorString);
-  [_connector adapter:self didFailAd:error];
-}
-
-- (void)getNativeAdWithAdTypes:(NSArray<GADAdLoaderAdType> *)adTypes
-                       options:(NSArray<GADAdLoaderOptions *> *)options {
-  id<GADMAdNetworkConnector> strongConnector = _connector;
-  // Validate adTypes.
-  if (![adTypes containsObject:kGADAdLoaderAdTypeUnifiedNative]) {
-    NSError *error = GADMAdapterIMobileErrorWithCodeAndDescription(
-        kGADErrorInvalidRequest, @"GADMediationAdapterIMobile only supports UnifiedNative.");
-    [strongConnector adapter:self didFailAd:error];
-    return;
-  }
-
-  // Call i-mobile SDK.
-  [ImobileSdkAds getNativeAdData:_spotID
-                            View:_sdkView
-                          Params:[[ImobileSdkAdsNativeParams alloc] init]
-                        Delegate:self];
-}
-
-#pragma mark - IMobileSdkAdsDelegate
-
-- (void)imobileSdkAdsSpot:(NSString *)spotId didFailWithValue:(ImobileSdkAdsFailResult)value {
-  [self stopBeingDelegate];
-  NSInteger errorCode = GADMAdapterIMobileAdMobErrorFromIMobileResult(value);
-  NSString *errorString = [NSString stringWithFormat:@"Failed to get an ad for spotID: %@", spotId];
-  NSError *error = GADMAdapterIMobileErrorWithCodeAndDescription(errorCode, errorString);
-  [_connector adapter:self didFailAd:error];
-}
-
-- (void)onNativeAdDataReciveCompleted:(NSString *)spotId nativeArray:(NSArray<id> *)nativeArray {
-  // Check ad data.
-  if ([nativeArray count] == 0) {
-    NSError *error =
-        GADMAdapterIMobileErrorWithCodeAndDescription(kGADErrorNoFill, @"No ads to show.");
-    [_connector adapter:self didFailAd:error];
-    return;
-  }
-
-  // Get ad image.
-  ImobileSdkAdsNativeObject *iMobileNativeAd = nativeArray[0];
-  [iMobileNativeAd getAdImageCompleteHandler:^(UIImage *image) {
-    id<GADMAdNetworkConnector> strongConnector = self->_connector;
-    if (!image) {
-      NSError *error = GADMAdapterIMobileErrorWithCodeAndDescription(
-          kGADErrorNoFill, @"Can't download native ad assets.");
-      [strongConnector adapter:self didFailAd:error];
-      return;
-    }
-    GADIMobileMediatedUnifiedNativeAd *unifiedAd =
-        [[GADIMobileMediatedUnifiedNativeAd alloc] initWithIMobileNativeAd:iMobileNativeAd
-                                                                     image:image];
-    [strongConnector adapter:self didReceiveMediatedUnifiedNativeAd:unifiedAd];
-  }];
+- (void)loadNativeAdForAdConfiguration:(GADMediationNativeAdConfiguration *)adConfiguration
+                     completionHandler:(GADMediationNativeLoadCompletionHandler)completionHandler {
+  _unifiedNativeAd =
+      [[GADMAdapterIMobileUnifiedNativeAd alloc] initWithAdConfiguration:adConfiguration];
+  [_unifiedNativeAd loadNativeAdWithCompletionHandler:completionHandler];
 }
 
 @end