+++ adapters/I-Mobile/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## i-mobile iOS Mediation Adapter Changelog
 
+#### Next Version
+- Added standardized adapter error codes and messages.
+- Adapter now limits requesting of multiple interstitial ads using the same i-mobile Spot ID.
+
 #### Version 2.0.33.0
 - Verified compatibility with i-mobile SDK version 2.0.33.
 - Now requires Google Mobile Ads SDK version 7.65.0 or higher.